// server/@types/make-recall-decision-api/models/RecallTypeSelectedValue.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type RecallTypeSelectedValue = {
    value?: RecallTypeSelectedValue.value;
    details?: string;
};

export namespace RecallTypeSelectedValue {

    export enum value {
        STANDARD = 'STANDARD',
        FIXED_TERM = 'FIXED_TERM',
        NO_RECALL = 'NO_RECALL',
    }


}